mqtt-test/pyproject.toml
[project]
name = "mqtt-test"
version = "0.1.0"
description = "Simple python script to test if mqtt works"
requires-python = ">=3.12"
dependencies = [
    "paho-mqtt>=2.1.0",
]
